Move code generation surface to gnostic top-level.
