+++ plugins/gnostic-go-generator/render_types.go
@@ -15,7 +15,7 @@
 package main
 
 import (
-	surface "github.com/googleapis/gnostic/plugins/gnostic-go-generator/surface"
+	surface "github.com/googleapis/gnostic/surface"
 )
 
 func (renderer *Renderer) RenderTypes() ([]byte, error) {
